obmc-console: Fix bugprone-narrowing-conversions
For example:

```
/mnt/host/andrew/home/andrew/src/openbmc/obmc-console/build/../console-server.c:769:9: error: narrowing conversion from 'ssize_t' (aka 'long') to signed type 'int' is implementation-defined [bugprone-narrowing-conversions,-warnings-as-errors]
                        rc = read(console->tty_fd, buf, sizeof(buf));
                             ^
```

+++ ringbuffer.c
@@ -125,7 +125,7 @@ static int ringbuffer_consumer_ensure_space(struct ringbuffer_consumer *rbc,
 					    size_t len)
 {
 	enum ringbuffer_poll_ret prc;
-	int force_len;
+	size_t force_len;
 
 	if (ringbuffer_space(rbc) >= len)
 		return 0;
